@@ -1,6 +1,6 @@
 import { BrowserModule } from '@angular/platform-browser';
 import { NgModule } from '@angular/core';
-import { HttpModule } from '@angular/http';
+import { HttpClientModule } from '@angular/common/http';
 
 import { AppComponent } from './app.component';
 import { MutantListComponent } from './mutant-list/mutant-list.component';
@@ -13,7 +13,7 @@ import { MutantListService } from './mutant-list/mutant-list.service';
   ],
   imports: [
     BrowserModule,
-    HttpModule
+    HttpClientModule
   ],
   providers: [MutantListService],
   bootstrap: [AppComponent]
